+++ src/gallery-sm-menu/js/menu-base.js
@@ -54,7 +54,7 @@ Fired when a menu item is shown.
 **/
 var EVT_SHOW = 'show';
 
-var MenuBase = Y.Base.create('menuBase', Y.Tree, [Y.Tree.Openable], {
+var MenuBase = Y.Base.create('menuBase', Y.Tree, [Y.Tree.Labelable, Y.Tree.Openable], {
     nodeClass: Y.Menu.Item,
 
     // -- Lifecycle ------------------------------------------------------------
@@ -93,7 +93,7 @@ var MenuBase = Y.Base.create('menuBase', Y.Tree, [Y.Tree.Openable], {
     **/
     disableItem: function (item, options) {
         if (!item.isDisabled()) {
-            this._fire(EVT_DISABLE, {item: item}, {
+            this._fireTreeEvent(EVT_DISABLE, {item: item}, {
                 defaultFn: this._defDisableFn,
                 silent   : options && options.silent
             });
@@ -114,7 +114,7 @@ var MenuBase = Y.Base.create('menuBase', Y.Tree, [Y.Tree.Openable], {
     **/
     enableItem: function (item, options) {
         if (item.isDisabled()) {
-            this._fire(EVT_ENABLE, {item: item}, {
+            this._fireTreeEvent(EVT_ENABLE, {item: item}, {
                 defaultFn: this._defEnableFn,
                 silent   : options && options.silent
             });
@@ -135,7 +135,7 @@ var MenuBase = Y.Base.create('menuBase', Y.Tree, [Y.Tree.Openable], {
     **/
     hideItem: function (item, options) {
         if (!item.isHidden()) {
-            this._fire(EVT_HIDE, {item: item}, {
+            this._fireTreeEvent(EVT_HIDE, {item: item}, {
                 defaultFn: this._defHideFn,
                 silent   : options && options.silent
             });
@@ -156,7 +156,7 @@ var MenuBase = Y.Base.create('menuBase', Y.Tree, [Y.Tree.Openable], {
     **/
     showItem: function (item, options) {
         if (item.isHidden()) {
-            this._fire(EVT_SHOW, {item: item}, {
+            this._fireTreeEvent(EVT_SHOW, {item: item}, {
                 defaultFn: this._defShowFn,
                 silent   : options && options.silent
             });