+++ src/gallery-sm-menu/js/menu-item.js
@@ -89,25 +89,29 @@ Y.extend(MenuItem, Y.Tree.Node, {
     },
 
     /**
-    Returns `true` if this menu item is currently disabled.
+    Returns `true` if this menu item or one of its ancestors is currently
+    disabled.
 
     @method isDisabled
-    @return {Boolean} `true` if this menu item is currently disabled, `false`
-        otherwise.
+    @return {Boolean} `true` if this menu item or one of its ancestors is
+        currently disabled, `false` otherwise.
     **/
     isDisabled: function () {
-        return !!this.state.disabled;
+        return !!this.state.disabled ||
+            (this.parent ? this.parent.isDisabled() : false);
     },
 
     /**
-    Returns `true` if this menu item is currently hidden.
+    Returns `true` if this menu item or one of its ancestors is currently
+    hidden.
 
     @method isHidden
-    @return {Boolean} `true` if this menu item is currently hidden, `false`
-        otherwise.
+    @return {Boolean} `true` if this menu item or one of its ancestors is
+        currently hidden, `false` otherwise.
     **/
     isHidden: function () {
-        return !!this.state.hidden;
+        return !!this.state.hidden ||
+            (this.parent ? this.parent.isHidden() : false);
     },
 
     /**