+++ src/gallery-sm-menu/HISTORY.md
@@ -1,4 +1,15 @@
 SmugMug Menu History
 ====================
 
-Nothing yet!
+## git
+
+* Now using `Y.Tree` from YUI core, which means Menu requires YUI 3.9.0+.
+
+* Y.Menu.Item's `isDisabled()` and `isHidden()` methods now return `true` if any
+  ancestor of the item is disabled or hidden, respectively.
+
+* Fixed: Menus take up space even when they're hidden. [#8]
+
+## 2012-02-07
+
+* Initial release.